Extract HTTP client logic to HttpClient and JavaHttpClient. Make HttpClient configurable.
